refactor: extracted `accountdb` from `condenserdb`
