+++ modules/db/mimic/condenserdb/condenser.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"log/slog"
 	"mimic/lib/utils"
-	"mimic/mock"
 	"mimic/modules/db"
 	"mimic/modules/db/mimic"
 	"time"
@@ -17,19 +16,17 @@ import (
 )
 
 type Condenser struct {
-	accounts     *mongo.Collection
 	orders       *mongo.Collection
 	transactions *mongo.Collection
 }
 
-var condenserDb = &Condenser{nil, nil, nil}
+var condenserDb = &Condenser{nil, nil}
 
 func Collection() *Condenser {
 	return condenserDb
 }
 
 func New(d *mimic.MimicDb) *Condenser {
-	condenserDb.accounts = db.NewCollection(d.DbInstance, "accounts")
 	condenserDb.orders = db.NewCollection(d.DbInstance, "orders")
 	condenserDb.transactions = db.NewCollection(d.DbInstance, "transactions")
 	return condenserDb
@@ -39,11 +36,6 @@ func New(d *mimic.MimicDb) *Condenser {
 func (c *Condenser) Init() error {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 	defer cancel()
-
-	db.CreateIndex(ctx, c.accounts, mongo.IndexModel{
-		Keys:    bson.D{{Key: "name", Value: 1}},
-		Options: options.Index().SetUnique(true).SetName("name_unique"),
-	})
 
 	db.CreateIndex(ctx, c.orders, mongo.IndexModel{
 		Keys:    bson.D{{Key: "id", Value: 1}},
@@ -62,27 +54,6 @@ func (c *Condenser) Init() error {
 }
 
 func (c *Condenser) Start() *promise.Promise[any] {
-	data := make(map[string]Account)
-	err := mock.GetMockData(&data, "condenser_api.get_accounts")
-	if err != nil {
-		panic(err)
-	}
-
-	entries := make([]any, 0, len(data))
-	for _, v := range data {
-		entries = append(entries, v)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
-	defer cancel()
-
-	result, err := c.accounts.InsertMany(ctx, entries)
-	if err != nil && !mongo.IsDuplicateKeyError(err) {
-		slog.Error("Failed to seed collection.", "err", err)
-	} else {
-		slog.Debug("Seed collection.", "collection", c.accounts.Name(), "new-record", len(result.InsertedIDs))
-	}
-
 	return utils.PromiseResolve[any](nil)
 }
 
@@ -92,28 +63,14 @@ func (c *Condenser) Stop() error {
 
 // Read Queries
 
-func (c *Condenser) QueryGetAccounts(
-	ctx context.Context,
-	accounts *[]Account,
-	namedQueries []string,
-) error {
-	filter := bson.M{"name": bson.M{"$in": namedQueries}}
-
-	cursor, err := c.accounts.Find(ctx, filter)
-	if err != nil {
-		return err
-	}
-
-	defer cursor.Close(ctx)
-
-	return cursor.All(ctx, accounts)
-}
-
 // Write Queries
 
 // Save `trx` to database, then return the transaction number, and writes the
 // inserted id to `trx`
-func (c *Condenser) NewTransaction(ctx context.Context, trx *Transaction) (int64, error) {
+func (c *Condenser) NewTransaction(
+	ctx context.Context,
+	trx *Transaction,
+) (int64, error) {
 	// write to db
 	result, err := c.transactions.InsertOne(ctx, trx)
 	if err != nil {
